# ci: configure e2e target properly in order to make fix affected e2e target execution on ci