--- a/change/@fluentui-react-f5fd080f-c867-41c4-be33-6b4e115a92ce.json
+++ b/change/@fluentui-react-f5fd080f-c867-41c4-be33-6b4e115a92ce.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "fix: configure e2e target properly in order to make fix affected e2e target execution on ci",
+  "packageName": "@fluentui/react",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}