--- a/change/@fluentui-web-components-6b58d04a-1f48-4aa3-ba16-e2ea9b0201de.json
+++ b/change/@fluentui-web-components-6b58d04a-1f48-4aa3-ba16-e2ea9b0201de.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "fix: configure e2e target properly in order to make fix affected e2e target execution on ci",
+  "packageName": "@fluentui/web-components",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}