--- a/nx.json
+++ b/nx.json
@@ -26,6 +26,11 @@
       "inputs": ["production", "^production"],
       "cache": true
     },
+    "build-storybook": {
+      "dependsOn": [],
+      "inputs": ["default", "{workspaceRoot}/.storybook/**", "{projectRoot}/.storybook/**"],
+      "cache": true
+    },
     "generate-api": {
       "dependsOn": ["^generate-api"],
       "cache": true
@@ -45,6 +50,11 @@
     },
     "test-ssr": {
       "cache": true
+    },
+    "e2e": {
+      "dependsOn": [],
+      "cache": true,
+      "inputs": ["default", "{projectRoot}/cypress.config.ts", "!{projectRoot}/**/?(*.)+cy.[jt]s?(x)?"]
     },
     "lint": {
       "dependsOn": ["build"],