--- a/packages/web-components/project.json
+++ b/packages/web-components/project.json
@@ -3,5 +3,8 @@
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "implicitDependencies": [],
-  "tags": ["platform:web", "web-components"]
+  "tags": ["platform:web", "web-components"],
+  "targets": {
+    "e2e": { "dependsOn": ["build-storybook"] }
+  }
 }