… packages/react/package.json