<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

every PR CI run triggers all/non deterministic projects `e2e` and `build-storybook` target = `--since` doesn't work

**Reproduction:**

Following PR afffects only react-nav-preview

https://github.com/microsoft/fluentui/pull/31671

thought all  projects that have `e2e` target are run 🐌 🐌 🐌 🐌 

<img width="1203" alt="image" src="https://github.com/microsoft/fluentui/assets/1223799/b72aa4c4-b009-4512-9656-fed247c6c408">


## New Behavior

PR ci run triggers `e2e` and `build-storybook` target only on affected projects = `--since` works

**Pipeline Perf boost:**

> PR containing package v9 changes

| target | Before | After |
|--------|--------|--------|
| e2e | 399s | 20s / **𝚫95% faster 🚅** |




## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Fixes partially https://github.com/microsoft/fluentui/issues/27359